Add `toArray` for HTMLCollection (#109)
* Add `toArray` for HTMLCollection

With a lot of help from @garyb, here's the `toArray` for HTMLCollection.
It returns the elements within the collection as an Array of Element
objects. It was suggested that this be returned as an Unfoldable, but
the `item` function is effectful, so the plan was ruined.

* Update toArray to use FFI

It's not great, but it's going to be more performant and just as
"correct" (for some definition of correctness).